+++ package.json
@@ -40,6 +40,8 @@
   },
   "dependencies": {
     "@bit/mui-org.material-ui.swipeable-drawer": "^4.9.10",
+    "@emotion/react": "^11.7.0",
+    "@emotion/styled": "^11.6.0",
     "@react-native-async-storage/async-storage": "^1.13.0",
     "@react-native-community/hooks": "^2.6.0",
     "@types/turndown": "^5.0.1",
@@ -102,7 +104,6 @@
     "redux-thunk": "^2.3.0",
     "reselect": "^4.0.0",
     "sanitize-html": "^2.4.0",
-    "styled-components": "^5.3.1",
     "tailwindcss": "^2.2.9",
     "truncate-html": "^1.0.3",
     "ts-key-enum": "^2.0.6",
@@ -145,7 +146,6 @@
     "@types/redux-mock-store": "^1.0.2",
     "@types/sanitize-html": "^2.3.1",
     "@types/sinonjs__fake-timers": "^6.0.2",
-    "@types/styled-components": "^5.1.13",
     "@types/styled-components-react-native": "^5.1.1",
     "@types/uuid": "^8.3.0",
     "@types/yallist": "^4.0.1",
