+++ src/App.css
@@ -927,10 +927,6 @@ h1 .num-contexts {
   margin-left: -12px;
 }
 
-.mobile .child > .thought-container > .bullet.show-contexts {
-  margin-right: -1.5px;
-}
-
 .archive-check {
   margin-left: -12px;
   margin-top: 5px;
@@ -2625,10 +2621,6 @@ code,
   background-color: black;
 }
 
-.invalid-option {
-  color: tomato !important;
-}
-
 .note {
   margin-top: 0;
   margin-bottom: 0;
@@ -2814,19 +2806,6 @@ code,
   to {
     color: white;
   }
-}
-
-.gray {
-  color: #666;
-}
-
-.graypulse {
-  color: #666;
-  -webkit-animation: toblack 400ms infinite alternate ease-in-out;
-}
-
-.dark .graypulse {
-  -webkit-animation: towhite 400ms infinite alternate ease-in-out;
 }
 
 .no-select {
