+++ src/components/Content.tsx
@@ -16,8 +16,7 @@ import Editable from './Editable'
 import { SimplePath, State } from '../@types'
 import { storage } from '../util/storage'
 import * as selection from '../device/selection'
-import styled from 'styled-components'
-import tw from 'twin.macro'
+import tw, { styled } from 'twin.macro'
 
 const tutorialLocal = storage.getItem('Settings/Tutorial') === 'On'
 const tutorialStepLocal = +(storage.getItem('Settings/Tutorial Step') || 1)
