+++ src/components/Input.tsx
@@ -1,5 +1,4 @@
-import styled from 'styled-components'
-import tw from 'twin.macro'
+import tw, { styled } from 'twin.macro'
 
 const variantStyles = {
   normal: tw`
