+++ src/components/TextLink.tsx
@@ -1,6 +1,5 @@
 import React, { FC } from 'react'
-import styled from 'styled-components'
-import tw from 'twin.macro'
+import tw, { styled } from 'twin.macro'
 
 /**
  * Text Link.
