+++ src/components/Thought/styles.tsx
@@ -1,0 +1,3 @@
+import { styled } from 'twin.macro'
+
+export const ThoughtComponentWrapper = styled.li``
